feat(ags): working hyprland/ags simple bar
In `ags/` the config of the bar is.
This is just the example config from the docs.

It required some changes to the Hyprland setup and the GTK setup.

Hyprland:

Changed to loading Hyprland from the flake (describes at Hyprland docs).
Split out hyprland config to `nixos/hyprland.nix`.
Here we load a lot of gnome utilities and start services, that are
needed.

GTK:

Split out the GTK theming into a separate file
`home-manager/gtk-theme.nix`